+++ Countly.h
@@ -13,8 +13,9 @@
 #import "CountlyFeedbackWidget.h"
 #import "CountlyViewTracking.h"
 #import "CountlyContentBuilder.h"
+#import "CountlyFeedbacks.h"
 #import "Resettable.h"
-#if (TARGET_OS_IOS || TARGET_OS_OSX)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_OSX )
 #import <UserNotifications/UserNotifications.h>
 #endif
 
@@ -376,7 +377,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 
 #pragma mark - Push Notification
-#if (TARGET_OS_IOS || TARGET_OS_OSX)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_OSX )
 #ifndef COUNTLY_EXCLUDE_PUSHNOTIFICATIONS
 /**
  * Shows default system dialog that asks for user's permission to display notifications.
@@ -552,7 +553,7 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)recordView:(NSString *)viewName segmentation:(NSDictionary<NSString *, id> *)segmentation DEPRECATED_MSG_ATTRIBUTE("Use '[views startView/startAutoStoppedView:]' method instead!");
 
-#if (TARGET_OS_IOS || TARGET_OS_TV)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_TV )
 /**
  * Adds exception for AutoViewTracking.
  * @discussion @c UIViewControllers with specified title or class name will be ignored by AutoViewTracking and their appearances and disappearances will not be recorded.
@@ -593,7 +594,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 
 #pragma mark - Feedbacks
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION )
 /**
  * Shows star-rating dialog manually and executes completion block after user's action.
  * @discussion Completion block has a single NSInteger parameter that indicates 1 to 5 star-rating given by user.
@@ -676,8 +677,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @discussion - Current device ID is @c CLYTemporaryDeviceID.
  * @param completionHandler A completion handler block to be executed when list is fetched successfully or there is an error.
  */
-- (void)getFeedbackWidgets:(void (^)(NSArray <CountlyFeedbackWidget *> * __nullable feedbackWidgets, NSError * __nullable error))completionHandler;
-
+- (void)getFeedbackWidgets:(void (^)(NSArray <CountlyFeedbackWidget *> * __nullable feedbackWidgets, NSError * __nullable error))completionHandler DEPRECATED_MSG_ATTRIBUTE("Use '[feedback getAvailableFeedbackWidgets:]' method instead!");
 
 /**
  * This is an experimental feature and it can have breaking changes
@@ -686,6 +686,11 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (CountlyContentBuilder *_Nonnull) content;
 
+/**
+ * Interface variable to access feedback widget functionalities.
+ * @discussion Feedback widget  interface for developer to interact with SDK.
+ */
+- (CountlyFeedbacks *) feedback;
 #endif