+++ CountlyConfig.h
@@ -13,16 +13,24 @@
 #import "CountlyExperimentalConfig.h"
 #import "CountlyContentConfig.h"
 
-#if (TARGET_OS_IOS || TARGET_OS_TV)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_TV )
 #import <UIKit/UIKit.h>
 #endif
 
 
 NS_ASSUME_NONNULL_BEGIN
 
+typedef enum : NSUInteger
+{
+    WIDGET_APPEARED,
+    WIDGET_CLOSED,
+} WidgetState;
+
+typedef void (^WidgetCallback)(WidgetState widgetState);
+
 //NOTE: Countly features
 typedef NSString* CLYFeature NS_EXTENSIBLE_STRING_ENUM;
-#if (TARGET_OS_IOS)
+#if (TARGET_OS_IOS || TARGET_OS_VISION )
 #ifndef COUNTLY_EXCLUDE_PUSHNOTIFICATIONS
 extern CLYFeature const CLYPushNotifications;
 #endif
@@ -218,7 +226,7 @@ typedef enum : NSUInteger
 
 #pragma mark -
 
-#if (TARGET_OS_IOS || TARGET_OS_TV)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_TV)
 /**
  * For enabling automatic view tacking.
  * @discussion If set, views will automatically track.