+++ CountlyPushNotifications.h
@@ -18,7 +18,7 @@ extern NSString* const kCountlyReservedEventPushAction;
 
 + (instancetype)sharedInstance;
 
-#if (TARGET_OS_IOS || TARGET_OS_OSX)
+#if (TARGET_OS_IOS || TARGET_OS_VISION || TARGET_OS_OSX)
 - (void)startPushNotifications;
 - (void)stopPushNotifications;
 - (void)askForNotificationPermissionWithOptions:(NSUInteger)options completionHandler:(void (^)(BOOL granted, NSError * error))completionHandler;