+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 * Mitigated an issue where an internal timer was not reset when going to foreground for `autoStoppedViews`
 * Mitigated an issue for `autoStoppedViews` could have not started when multiple views were open at the same time while going to foreground
 
+## 24.7.4
+* Added visionOS build support
+* Added `CountlyFeedbacks:` interface with new view methods (Access with `Countly.sharedInstance.feedback`):
+    * Method to present feedback widget (wih an optional widget selector(name, ID or tag) string and a Callback):
+        * `presentNPS`
+        * `presentSurvey`
+        * `presentRating`
+    * `getAvailableFeedbackWidgets` method to retrieve available feedback widgets with a completion handler.
+
+* Mitigated an issue with the feedback widget URL encoding on iOS 16 and earlier, which prevented the widget from displaying
+* Mitigated an issue with content fetch URL encoding on iOS 16 and earlier, which caused the request to fail
+    
+* Deprecated `getFeedbackWidgets` method, you should use `[feedback getAvailableFeedbackWidgets:]` method instead
+
 ## 24.7.3
 * Added current view names to event segmentation based on the `enablePreviousNameRecording` (Experimental!)
 * Updated the SDK to ensure compatibility with the latest server response models